requirements.txt
Django>=3.2,<4.0
django-phonenumber-field
djangorestframework
django-cors-headers
mysqlclient
phonenumbers